release: 8.0.7 — code-strip sweep
Remove 10 items of dead, redundant, or parallel-surface code.
No new features.

A1  Delete `mdds/normalize.rs` forwarding layer. Wire canonicalizers stay
    in `crate::wire_semantics`; MDDS-scoped `normalize_interval`,
    `normalize_time_of_day`, and `contract_spec!` move to `mdds/endpoints.rs`.
A2  Delete `fpss::session::reconnect` (90 LOC, zero callers).
A3  Delete `crate::right` re-export shim; `parse_right` / `parse_right_strict`
    / `ParsedRight` now resolve via `pub use tdbe::right::*`.
A4  Delete `crate::retry`. `StatusClass` moves into `macros.rs` as private;
    the unreachable `RetryOutcome` / `classify_status` / `retry_transient`
    trio and the `#![allow(dead_code)]` crate-header attribute are gone.
A5  Fold `tdbe::errors` into `tdbe::error`. `HTTP_STATUS_CODE_KEY` +
    `error_from_http_code` reachable at the new path; `error_name` helper
    deleted. `tdbe` bumps to 0.12.0 for the module-removal breaking change.
A6  Delete 24 `subscribe_*_stock` / `subscribe_*_option` / `unsubscribe_*`
    shortcut methods on `FpssClient` + `ThetaDataDx`. Callers use the
    `Contract`-taking methods directly.
A8  Narrow `pub mod unified` and `pub mod registry` to `pub(crate)`; keep
    documented types public via `pub use`. Narrow `DirectConfig::production_defaults`
    to `pub(crate)`. Delete `pub use prost` and the unused `MddsClient::raw_query`,
    `raw_query_info`, `channel` methods. `sdks/python` now depends on `prost`
    directly (`=0.14.3`).
A9  Delete 61 plain-name `tdx_<endpoint>` FFI entry points, the four
    `ffi_list_endpoint!` / `ffi_typed_endpoint!` / `ffi_typed_snapshot_endpoint!` /
    `ffi_list_endpoint_no_params!` macros, and the hand-written
    `ffi/src/historical.rs`. Every SDK already calls `_with_options`.
    `sdks/cpp/include/thetadx.h` loses the 61 legacy declarations;
    `ffi/src/endpoints.rs` is the new `include!` seam for the generated
    `endpoint_request_options.rs` + `endpoint_with_options.rs`.
A10 Delete 61 `MddsClient::<endpoint>_with_deadline` sibling methods. SDK
    generators now wrap the bare call in `tokio::time::timeout` locally.

Version bumps: every Rust crate `8.0.6 → 8.0.7`; `tdbe` `0.11.0 → 0.12.0`;
all five `Cargo.lock`s refreshed; `sdks/typescript/package.json` + every
npm platform subpackage bumped to 8.0.7.

Documentation handoffs folded in: 4 doc-comment path refs updated from
`build_support/endpoints.rs` to `build_support/endpoints/`; `FpssData::Quote`
/ `FpssData::Trade` doc example in `fpss/mod.rs` updated to the v8 `contract`
field shape; `sdks/python/src/lib.rs` + `sdks/typescript/src/lib.rs` header
rephrased factually.

Verification (all green):
- `cargo fmt --all -- --check`
- `cargo clippy --workspace --all-targets -- -D warnings`
- `cargo clippy --workspace --all-targets --features "polars,arrow" -- -D warnings`
- `cargo test --workspace --lib` (256 passed)
- `cargo test --workspace --features "polars,arrow" --lib`
- `bash crates/thetadatadx/tests/regen_byte_identical.sh` (56 files byte-identical)
- `cargo run -p thetadatadx --features config-file --bin generate_sdk_surfaces -- --check`
