+++ sdks/cpp/include/thetadx.h
@@ -448,299 +448,6 @@ void tdx_client_free(TdxClient* client);
 /** Free a string returned by any tdx_* function. */
 void tdx_string_free(char* s);
 
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Stock — List endpoints (2)                                            */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 1. List all stock symbols. Returns TdxStringArray. */
-TdxStringArray tdx_stock_list_symbols(const TdxClient* client);
-
-/** 2. List available dates for a stock by request type. Returns TdxStringArray. */
-TdxStringArray tdx_stock_list_dates(const TdxClient* client, const char* request_type, const char* symbol);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Stock — Snapshot endpoints (4)                                        */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 3. Get latest OHLC snapshot. symbols is a C array of C strings with length symbols_len. */
-TdxOhlcTickArray tdx_stock_snapshot_ohlc(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/** 4. Get latest trade snapshot. symbols is a C array of C strings with length symbols_len. */
-TdxTradeTickArray tdx_stock_snapshot_trade(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/** 5. Get latest NBBO quote snapshot. symbols is a C array of C strings with length symbols_len. */
-TdxQuoteTickArray tdx_stock_snapshot_quote(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/** 6. Get latest market value snapshot. symbols is a C array of C strings with length symbols_len. */
-TdxMarketValueTickArray tdx_stock_snapshot_market_value(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Stock — History endpoints (5 + bonus)                                 */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 7. Fetch stock EOD history. */
-TdxEodTickArray tdx_stock_history_eod(const TdxClient* client, const char* symbol,
-                                      const char* start_date, const char* end_date);
-
-/** 8. Fetch stock intraday OHLC. */
-TdxOhlcTickArray tdx_stock_history_ohlc(const TdxClient* client, const char* symbol,
-                                        const char* date, const char* interval);
-
-/** 8b. Fetch stock OHLC across date range. */
-TdxOhlcTickArray tdx_stock_history_ohlc_range(const TdxClient* client, const char* symbol,
-                                              const char* start_date, const char* end_date,
-                                              const char* interval);
-
-/** 9. Fetch all trades on a date. */
-TdxTradeTickArray tdx_stock_history_trade(const TdxClient* client, const char* symbol, const char* date);
-
-/** 10. Fetch NBBO quotes. */
-TdxQuoteTickArray tdx_stock_history_quote(const TdxClient* client, const char* symbol,
-                                          const char* date, const char* interval);
-
-/** 11. Fetch combined trade + quote ticks. */
-TdxTradeQuoteTickArray tdx_stock_history_trade_quote(const TdxClient* client, const char* symbol, const char* date);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Stock — At-Time endpoints (2)                                         */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 12. Fetch trade at a specific time across a date range. */
-TdxTradeTickArray tdx_stock_at_time_trade(const TdxClient* client, const char* symbol,
-                                          const char* start_date, const char* end_date,
-                                          const char* time_of_day);
-
-/** 13. Fetch quote at a specific time across a date range. */
-TdxQuoteTickArray tdx_stock_at_time_quote(const TdxClient* client, const char* symbol,
-                                          const char* start_date, const char* end_date,
-                                          const char* time_of_day);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Option — List endpoints (5)                                           */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 14. List all option underlyings. */
-TdxStringArray tdx_option_list_symbols(const TdxClient* client);
-
-/** 15. List available dates for an option contract. */
-TdxStringArray tdx_option_list_dates(const TdxClient* client, const char* request_type,
-                                     const char* symbol, const char* expiration,
-                                     const char* strike, const char* right);
-
-/** 16. List expiration dates. */
-TdxStringArray tdx_option_list_expirations(const TdxClient* client, const char* symbol);
-
-/** 17. List strike prices. */
-TdxStringArray tdx_option_list_strikes(const TdxClient* client, const char* symbol,
-                                       const char* expiration);
-
-/** 18. List all option contracts on a date. */
-TdxOptionContractArray tdx_option_list_contracts(const TdxClient* client, const char* request_type,
-                                                 const char* symbol, const char* date);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Option — Snapshot endpoints (10)                                      */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 19. Get latest OHLC snapshot for options. */
-TdxOhlcTickArray tdx_option_snapshot_ohlc(const TdxClient* client, const char* symbol,
-                                          const char* expiration, const char* strike, const char* right);
-
-/** 20. Get latest trade snapshot for options. */
-TdxTradeTickArray tdx_option_snapshot_trade(const TdxClient* client, const char* symbol,
-                                            const char* expiration, const char* strike, const char* right);
-
-/** 21. Get latest NBBO quote snapshot for options. */
-TdxQuoteTickArray tdx_option_snapshot_quote(const TdxClient* client, const char* symbol,
-                                            const char* expiration, const char* strike, const char* right);
-
-/** 22. Get latest open interest snapshot. */
-TdxOpenInterestTickArray tdx_option_snapshot_open_interest(const TdxClient* client, const char* symbol,
-                                                           const char* expiration, const char* strike, const char* right);
-
-/** 23. Get latest market value snapshot for options. */
-TdxMarketValueTickArray tdx_option_snapshot_market_value(const TdxClient* client, const char* symbol,
-                                                         const char* expiration, const char* strike, const char* right);
-
-/** 24. Get IV snapshot for options. */
-TdxIvTickArray tdx_option_snapshot_greeks_implied_volatility(const TdxClient* client, const char* symbol,
-                                                             const char* expiration, const char* strike, const char* right);
-
-/** 25. Get all Greeks snapshot. */
-TdxGreeksTickArray tdx_option_snapshot_greeks_all(const TdxClient* client, const char* symbol,
-                                                   const char* expiration, const char* strike, const char* right);
-
-/** 26. Get first-order Greeks snapshot. */
-TdxGreeksTickArray tdx_option_snapshot_greeks_first_order(const TdxClient* client, const char* symbol,
-                                                           const char* expiration, const char* strike, const char* right);
-
-/** 27. Get second-order Greeks snapshot. */
-TdxGreeksTickArray tdx_option_snapshot_greeks_second_order(const TdxClient* client, const char* symbol,
-                                                            const char* expiration, const char* strike, const char* right);
-
-/** 28. Get third-order Greeks snapshot. */
-TdxGreeksTickArray tdx_option_snapshot_greeks_third_order(const TdxClient* client, const char* symbol,
-                                                           const char* expiration, const char* strike, const char* right);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Option — History endpoints (6)                                        */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 29. Fetch EOD option data. */
-TdxEodTickArray tdx_option_history_eod(const TdxClient* client, const char* symbol, const char* expiration,
-                                       const char* strike, const char* right,
-                                       const char* start_date, const char* end_date);
-
-/** 30. Fetch intraday OHLC for options. */
-TdxOhlcTickArray tdx_option_history_ohlc(const TdxClient* client, const char* symbol, const char* expiration,
-                                         const char* strike, const char* right,
-                                         const char* date, const char* interval);
-
-/** 31. Fetch all trades for an option. */
-TdxTradeTickArray tdx_option_history_trade(const TdxClient* client, const char* symbol, const char* expiration,
-                                           const char* strike, const char* right, const char* date);
-
-/** 32. Fetch NBBO quotes for an option. */
-TdxQuoteTickArray tdx_option_history_quote(const TdxClient* client, const char* symbol, const char* expiration,
-                                           const char* strike, const char* right,
-                                           const char* date, const char* interval);
-
-/** 33. Fetch combined trade + quote for an option. */
-TdxTradeQuoteTickArray tdx_option_history_trade_quote(const TdxClient* client, const char* symbol, const char* expiration,
-                                                      const char* strike, const char* right, const char* date);
-
-/** 34. Fetch open interest history. */
-TdxOpenInterestTickArray tdx_option_history_open_interest(const TdxClient* client, const char* symbol, const char* expiration,
-                                                          const char* strike, const char* right, const char* date);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Option — History Greeks endpoints (11)                                */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 35-45: Greeks history endpoints all return TdxGreeksTickArray or TdxIvTickArray. */
-TdxGreeksTickArray tdx_option_history_greeks_eod(const TdxClient* client, const char* symbol, const char* expiration,
-                                                  const char* strike, const char* right,
-                                                  const char* start_date, const char* end_date);
-
-TdxGreeksTickArray tdx_option_history_greeks_all(const TdxClient* client, const char* symbol, const char* expiration,
-                                                  const char* strike, const char* right,
-                                                  const char* date, const char* interval);
-
-TdxGreeksTickArray tdx_option_history_trade_greeks_all(const TdxClient* client, const char* symbol, const char* expiration,
-                                                        const char* strike, const char* right, const char* date);
-
-TdxGreeksTickArray tdx_option_history_greeks_first_order(const TdxClient* client, const char* symbol, const char* expiration,
-                                                          const char* strike, const char* right,
-                                                          const char* date, const char* interval);
-
-TdxGreeksTickArray tdx_option_history_trade_greeks_first_order(const TdxClient* client, const char* symbol, const char* expiration,
-                                                                const char* strike, const char* right, const char* date);
-
-TdxGreeksTickArray tdx_option_history_greeks_second_order(const TdxClient* client, const char* symbol, const char* expiration,
-                                                           const char* strike, const char* right,
-                                                           const char* date, const char* interval);
-
-TdxGreeksTickArray tdx_option_history_trade_greeks_second_order(const TdxClient* client, const char* symbol, const char* expiration,
-                                                                 const char* strike, const char* right, const char* date);
-
-TdxGreeksTickArray tdx_option_history_greeks_third_order(const TdxClient* client, const char* symbol, const char* expiration,
-                                                          const char* strike, const char* right,
-                                                          const char* date, const char* interval);
-
-TdxGreeksTickArray tdx_option_history_trade_greeks_third_order(const TdxClient* client, const char* symbol, const char* expiration,
-                                                                const char* strike, const char* right, const char* date);
-
-TdxIvTickArray tdx_option_history_greeks_implied_volatility(const TdxClient* client, const char* symbol, const char* expiration,
-                                                             const char* strike, const char* right,
-                                                             const char* date, const char* interval);
-
-TdxIvTickArray tdx_option_history_trade_greeks_implied_volatility(const TdxClient* client, const char* symbol, const char* expiration,
-                                                                   const char* strike, const char* right, const char* date);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Option — At-Time endpoints (2)                                        */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 46. Fetch trade at a specific time for an option. */
-TdxTradeTickArray tdx_option_at_time_trade(const TdxClient* client, const char* symbol, const char* expiration,
-                                           const char* strike, const char* right,
-                                           const char* start_date, const char* end_date, const char* time_of_day);
-
-/** 47. Fetch quote at a specific time for an option. */
-TdxQuoteTickArray tdx_option_at_time_quote(const TdxClient* client, const char* symbol, const char* expiration,
-                                           const char* strike, const char* right,
-                                           const char* start_date, const char* end_date, const char* time_of_day);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Index — List endpoints (2)                                            */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 48. List all index symbols. */
-TdxStringArray tdx_index_list_symbols(const TdxClient* client);
-
-/** 49. List available dates for an index. */
-TdxStringArray tdx_index_list_dates(const TdxClient* client, const char* symbol);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Index — Snapshot endpoints (3)                                        */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 50. Get latest OHLC snapshot for indices. symbols is a C array of C strings with length symbols_len. */
-TdxOhlcTickArray tdx_index_snapshot_ohlc(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/** 51. Get latest price snapshot for indices. symbols is a C array of C strings with length symbols_len. */
-TdxPriceTickArray tdx_index_snapshot_price(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/** 52. Get latest market value for indices. symbols is a C array of C strings with length symbols_len. */
-TdxMarketValueTickArray tdx_index_snapshot_market_value(const TdxClient* client, const char* const* symbols, size_t symbols_len);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Index — History endpoints (3)                                         */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 53. Fetch EOD index data. */
-TdxEodTickArray tdx_index_history_eod(const TdxClient* client, const char* symbol,
-                                      const char* start_date, const char* end_date);
-
-/** 54. Fetch intraday OHLC for an index. */
-TdxOhlcTickArray tdx_index_history_ohlc(const TdxClient* client, const char* symbol,
-                                        const char* start_date, const char* end_date,
-                                        const char* interval);
-
-/** 55. Fetch intraday price history for an index. */
-TdxPriceTickArray tdx_index_history_price(const TdxClient* client, const char* symbol,
-                                          const char* date, const char* interval);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Index — At-Time endpoints (1)                                         */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 56. Fetch index price at a specific time. */
-TdxPriceTickArray tdx_index_at_time_price(const TdxClient* client, const char* symbol,
-                                          const char* start_date, const char* end_date,
-                                          const char* time_of_day);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Calendar endpoints (3)                                                */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 57. Check whether the market is open today. */
-TdxCalendarDayArray tdx_calendar_open_today(const TdxClient* client);
-
-/** 58. Get calendar for a specific date. */
-TdxCalendarDayArray tdx_calendar_on_date(const TdxClient* client, const char* date);
-
-/** 59. Get calendar for an entire year. */
-TdxCalendarDayArray tdx_calendar_year(const TdxClient* client, const char* year);
-
-/* ═══════════════════════════════════════════════════════════════════════ */
-/*  Interest Rate endpoints (1)                                           */
-/* ═══════════════════════════════════════════════════════════════════════ */
-
-/** 60. Fetch EOD interest rate history. */
-TdxInterestRateTickArray tdx_interest_rate_history_eod(const TdxClient* client, const char* symbol,
-                                                       const char* start_date, const char* end_date);
-
 /* Generated option-aware endpoint declarations. */
 #include "endpoint_with_options.h.inc"
