+++ sdks/python/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "thetadatadx-py"
-version = "8.0.6"
+version = "8.0.7"
 edition = "2021"
 description = "Python bindings for thetadatadx — native ThetaData SDK powered by Rust"
 license = "Apache-2.0"
@@ -24,7 +24,12 @@ crate-type = ["cdylib"]
 [dependencies]
 # The Rust SDK we're wrapping
 thetadatadx = { path = "../../crates/thetadatadx" }
-tdbe = { version = "0.11.0", path = "../../crates/tdbe" }
+tdbe = { version = "0.12.0", path = "../../crates/tdbe" }
+
+# Direct prost dep for decoding `thetadatadx::proto::ResponseData` bytes in
+# the `decode_response_bytes` hook. The main crate no longer re-exports
+# `prost`; we pull it in here at the same version pin.
+prost = "=0.14.3"
 
 # PyO3 bindings. We target the stable ABI so one wheel per platform supports
 # every Python version from 3.9 upward.
