+++ sdks/typescript/npm/darwin-arm64/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx-darwin-arm64",
-  "version": "8.0.6",
+  "version": "8.0.7",
   "os": [
     "darwin"
   ],
