+++ sdks/typescript/npm/linux-x64-gnu/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx-linux-x64-gnu",
-  "version": "8.0.6",
+  "version": "8.0.7",
   "os": [
     "linux"
   ],
