+++ sdks/typescript/npm/win32-x64-msvc/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx-win32-x64-msvc",
-  "version": "8.0.6",
+  "version": "8.0.7",
   "os": [
     "win32"
   ],
